Added song array to scanner class. Did not test
Need to figure out logic here for the scanner. Is the songs array
already existent and are we appending to it.  Determined based on need,
we should discuss.
Perhaps best to leave as is. If appending needed, we can call method,
return arraylist, and append as needed at that point.